Fixed long long build

+++ xtransmit/metrics_latency.cpp
@@ -14,7 +14,7 @@ using namespace chrono;
 void latency::submit_sample(const time_point& sample_time, const time_point& current_time)
 {
 	const auto delay = current_time - sample_time;
-	const int64_t delay_us = duration_cast<microseconds>(delay).count();
+	const long long delay_us = duration_cast<microseconds>(delay).count();
 	m_latency_max_us = max(m_latency_max_us, delay_us);
 	m_latency_min_us = min(m_latency_min_us, delay_us);
 	m_latency_avg_us = m_latency_avg_us == 0
